more work on darm...

+++ src/main.c
@@ -1,5 +1,6 @@
 #include "darm.h"
 
+#include <errno.h>
 #include <inttypes.h>
 
 #include <stdio.h>
@@ -21,7 +22,7 @@ int main (int argc, char * argv[])
     filesize = ftell(fh);
     fseek(fh, 0, SEEK_SET);
 
-    printf("filesize: %d\n", filesize);
+    printf("filesize: %d\n", (int) filesize);
 
     buf = (unsigned char *) malloc(filesize);
 
@@ -39,9 +40,9 @@ int main (int argc, char * argv[])
 
         struct _darm darm;
         if (darm_dis(&darm, ins)) {
-            printf("%04x %08x <invalid instruction>\n", i, ins);
+            printf("%04x %08x <invalid instruction>\n", (unsigned int) i, ins);
         } else {
-            printf("%04x %08x %s\n", i, ins, darm_str(&darm, i));
+            printf("%04x %08x %s\n", (unsigned int) i, ins, darm_str(&darm, i));
         }
     }
